+++ packages/ListBox/CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [1.4.0]
+
+- Fix some issue with keyboard functionality introduced when working with ListBoxWithTags on the Trigger component.
+- Remove dispatch toggle from Trigger and when with an ordinary if and else, adding clarity to whats happening.
+
 ## [1.0.1]
 
 - Fix accessiblity issue so options are now focusable and read to screen reader technology